rename: performance관련 타입 수정
